refactor Synth into Synth (graded base mostly), SynthLinearBase, and Common module
